Preparing for system test.
