+++ test/Nexus.Link.Libraries.Web.AspNet.Tests/InboundPipe/ValueTranslatorFilterSystemTests.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Nexus.Link.Libraries.Web.AspNet.Tests.InboundPipe
+{
+    class ValueTranslatorFilterSystemTests
+    {
+    }
+}
