+++ test/Nexus.Link.Libraries.Web.AspNet.Tests/InboundPipe/ValueTranslatorFilterTests.cs
@@ -1,9 +1,9 @@
 ﻿
 using Newtonsoft.Json.Linq;
+using Nexus.Link.Libraries.Web.AspNet.Tests.InboundPipe.Support;
 #if NETCOREAPP
 using System.Collections.Generic;
 using System.Threading;
-using Nexus.Link.Libraries.Core.Assert;
 using Nexus.Link.Libraries.Web.AspNet.Pipe.Inbound;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc.Controllers;
@@ -15,10 +15,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
 using Nexus.Link.Libraries.Core.Application;
-using Nexus.Link.Libraries.Core.Storage.Model;
 using Nexus.Link.Libraries.Core.Translation;
-using Nexus.Link.Libraries.Crud.Interfaces;
-#pragma warning disable 659
 
 namespace Nexus.Link.Libraries.Web.AspNet.Tests.InboundPipe
 {
@@ -86,40 +83,6 @@ namespace Nexus.Link.Libraries.Web.AspNet.Tests.InboundPipe
             var outFoo = jObject.ToObject<Foo>();
             Assert.IsNotNull(outFoo);
             Assert.AreEqual(outId, outFoo.Id);
-        }
-
-        [ApiController]
-        [Route("api/Foos")]
-        private class FoosController : ControllerBase, IUpdateAndReturn<Foo, string>
-        {
-            /// <inheritdoc />
-            [HttpPut]
-            [Route("{id}")]
-            public Task<Foo> UpdateAndReturnAsync([TranslationConcept("foo.id")]string id, Foo item,
-                CancellationToken token = default(CancellationToken))
-            {
-                InternalContract.RequireNotNullOrWhiteSpace(id, nameof(id));
-                InternalContract.RequireNotNull(item, nameof(item));
-                Assert.IsTrue(item.Id.StartsWith("(foo.id!"));
-                InternalContract.Require(id == item.Id, $"Expected {nameof(id)} to be identical to {nameof(item)}.{nameof(item.Id)}.");
-                return Task.FromResult(item);
-            }
-        }
-
-        private class Foo : IUniquelyIdentifiable<string>
-        {
-            /// <inheritdoc />
-            [TranslationConcept("foo.id")]
-            public string Id { get; set; }
-
-            public string Name { get; set; }
-
-            /// <inheritdoc />
-            public override bool Equals(object obj)
-            {
-                if (!(obj is Foo other)) return false;
-                return other.Id == Id && other.Name == Name;
-            }
         }
     }
 }
